requirements.txt
paho-mqtt>=2.1.0
PyYAML>=6.0
requests>=2.28.0
